= http: refactor auth headers - move Challenge to AdditionalRules - move Www-/Proxy-Authenticate to SimpleHeaders - move Proxy-Authorization to AuthorizationHeader - delete ProxyAuthenticateHeader, ProxyAuthorizationHeader, WwwAuthenticateHeader
