+++ spray-http/src/main/scala/spray/http/parser/AdditionalRules.scala
@@ -31,6 +31,13 @@ private[parser] trait AdditionalRules {
     Digit ~ optional(Digit ~ optional(Digit))
   }
 
+  def Challenge = rule {
+    AuthScheme ~ zeroOrMore(AuthParam, separator = ListSep) ~~> { (scheme, params) ⇒
+      val (realms, otherParams) = params.partition(_._1 == "realm")
+      HttpChallenge(scheme, realms.headOption.map(_._2).getOrElse(""), otherParams.toMap)
+    }
+  }
+
   def AuthScheme = rule {
     Token ~ OptWS
   }
