+++ spray-http/src/main/scala/spray/http/parser/HttpParser.scala
@@ -36,10 +36,7 @@ object HttpParser extends Parser with ProtocolParameterRules with AdditionalRule
     with ContentEncodingHeader
     with ContentTypeHeader
     with CookieHeaders
-    with ProxyAuthenticateHeader
-    with ProxyAuthorizationHeader
     with SimpleHeaders
-    with WwwAuthenticateHeader
     with CORSHeaders {
 
   // all string literals automatically receive a trailing optional whitespace
