--- spray-http/src/main/scala/spray/http/parser/ProxyAuthorizationHeader.scala
@@ -1,29 +1,0 @@
-/*
- * Copyright (C) 2011-2013 spray.io
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package spray
-package http
-package parser
-
-import org.parboiled.scala._
-
-private[parser] trait ProxyAuthorizationHeader extends AuthorizationHeader {
-  this: Parser with ProtocolParameterRules with AdditionalRules ⇒
-
-  def `*Proxy-Authorization` = rule {
-    CredentialDef ~ EOI ~~> (HttpHeaders.`Proxy-Authorization`(_))
-  }
-}
